+++ src/pages/Rop.jsx
@@ -1,6 +1,7 @@
 import {
   Edit,
   Edit2,
+  Plus,
   PlusCircle,
   PlusCircleIcon,
   RefreshCcw,
@@ -8,6 +9,17 @@ import {
   Trash,
 } from "lucide-react";
 import { Button } from "../components/ui/button";
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogTrigger,
+} from "@/components/ui/alert-dialog";
 import {
   Drawer,
   DrawerContent,
@@ -32,6 +44,7 @@ import {
   TooltipContent,
   TooltipTrigger,
 } from "../components/ui/tooltip";
+import { useLoadingBar } from "react-top-loading-bar";
 
 export default function Rop() {
   const { user } = useAppStore();
@@ -51,6 +64,10 @@ export default function Rop() {
   const [editLoading, setEditLoading] = useState(false);
   const [removeLoading, setRemoveLoading] = useState(false);
   const [companiesLoading, setCompaniesLoading] = useState(false);
+  const { start, complete } = useLoadingBar({
+    color: "#5ea500",
+    height: 3,
+  });
 
   // Permanently states
   const [deletingRop, setDeletingRop] = useState(null);
@@ -107,6 +124,7 @@ export default function Rop() {
 
   // Read
   async function get() {
+    start();
     let req;
     const token = JSON.parse(localStorage.getItem("user")).accessToken;
     setGetLoading(true);
@@ -134,6 +152,7 @@ export default function Rop() {
     }
 
     setGetLoading(false);
+    complete();
   }
 
   // Update
@@ -308,13 +327,8 @@ export default function Rop() {
   function handleDelete(id) {
     const foundRop = rops.find((rop) => rop.id === id);
     setDeletingRop(foundRop);
-    const check = confirm(
-      `Rostan ham <${foundRop.email}> ni o'chirib yubormoqchimisiz? Keyin bu operatsiyani orqaga qaytarib bo'lmaydi!`
-    );
 
-    if (check) {
-      remove(id);
-    }
+    remove(id);
   }
 
   function handleEdit(id) {
@@ -337,7 +351,7 @@ export default function Rop() {
   if (user) {
     if (getLoading) {
       return (
-        <div className="w-full h-full flex items-center justify-center">
+        <div className="w-full h-full flex items-center justify-center absolute bg-background z-50 inset-0">
           <div className="flex gap-4 items-center animate-pulse">
             <img
               className="w-20 h-20 rounded shadow"
@@ -411,20 +425,44 @@ export default function Rop() {
 
                       <Tooltip>
                         <TooltipTrigger asChild>
-                          <Button
-                            onClick={() => {
-                              handleDelete(id);
-                            }}
-                            disabled={deletingRop?.id === id && removeLoading}
-                            variant="destructive"
-                            size="icon-sm"
-                          >
-                            {deletingRop?.id === id && removeLoading ? (
-                              <RefreshCcw className="animate-spin" />
-                            ) : (
-                              <Trash />
-                            )}
-                          </Button>
+                          <AlertDialog>
+                            <AlertDialogTrigger asChild>
+                              <Button
+                                disabled={
+                                  deletingRop?.id === id && removeLoading
+                                }
+                                variant="destructive"
+                                size="icon-sm"
+                              >
+                                {deletingRop?.id === id && removeLoading ? (
+                                  <RefreshCcw className="animate-spin" />
+                                ) : (
+                                  <Trash />
+                                )}
+                              </Button>
+                            </AlertDialogTrigger>
+                            <AlertDialogContent>
+                              <AlertDialogHeader>
+                                <AlertDialogTitle>
+                                  Rostan ham o'chirib yubormoqchimisiz?
+                                </AlertDialogTitle>
+                                <AlertDialogDescription>
+                                  Keyin bu operatsiyani orqaga qaytarib
+                                  bo'lmaydi!
+                                </AlertDialogDescription>
+                              </AlertDialogHeader>
+                              <AlertDialogFooter>
+                                <AlertDialogCancel>Yo'q</AlertDialogCancel>
+                                <AlertDialogAction
+                                  onClick={() => {
+                                    handleDelete(id);
+                                  }}
+                                >
+                                  Ha
+                                </AlertDialogAction>
+                              </AlertDialogFooter>
+                            </AlertDialogContent>
+                          </AlertDialog>
                         </TooltipTrigger>
                         <TooltipContent>
                           <p>O'chirish</p>
@@ -439,14 +477,14 @@ export default function Rop() {
         ) : (
           <div className="w-full h-full flex justify-center items-center animate-fade-in">
             <div className="flex flex-col items-center text-center w-full max-w-sm">
-              <h3 className="text-2xl mb-3 font-medium">
-                Hali boshqaruvchi mavjud emas!
-              </h3>
-              <p className="text-muted-foreground mb-5">
-                Boshqaruvchi yaratishni istasangiz "Istayman" tugmasini bosing.
-              </p>
+              <img
+                className="w-50 object-center select-none mb-5"
+                src="/no-data.svg"
+                alt=""
+              />
+              <p className="mb-5">Hozircha ma'lumot yo'q</p>
               <Button onClick={handleAddModal} variant="secondary">
-                Istayman
+                <Plus /> Qo'shish
               </Button>
             </div>
           </div>
